+++ docs/changelog.rst
@@ -25,6 +25,8 @@ Bug fixes:
   matching. :bug:`5189`
 - :doc:`plugins/discogs` Fixed inconsistency in stripping disambiguation from
   artists but not labels. :bug:`5366`
+- :doc:`plugins/chroma` :doc:`plugins/bpsync` Fix plugin loading issue caused by
+  an import of another :class:`beets.plugins.BeetsPlugin` class. :bug:`6033`
 
 For packagers: